+++ src/app/components/frame/error-boundary/ErrorBoundary.style.tsx
@@ -11,7 +11,6 @@ export const ErrorBoundaryStyle = (theme: Theme) => ({
 		maxWidth: 500
 	},
 	sLink: {
-		cursor: 'pointer',
 		fontSize: theme.typography.pxToRem(17)
 	}
 });