+++ src/app/screens/business/robots/content/detail/general/RobotDetailGeneral.tsx
@@ -1,15 +1,17 @@
-import { Card, CardContent, Grid, Typography } from '@material-ui/core';
+import { Card, CardContent, Grid, IconButton, Typography } from '@material-ui/core';
 import { InfoOutlined } from '@material-ui/icons';
-import { FC } from 'react';
+import { Edit } from '@material-ui/icons';
+import { FC, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import Status from '../../../../../../components/common/status/Status';
 import Tooltip from '../../../../../../components/common/tooltip/Tooltip';
 import { AppConfigService } from '../../../../../../services';
 import { momentFormat1 } from '../../../../../../utilities/methods/Moment';
-import { strRemoveSymbols } from '../../../../../../utilities/methods/StringUtilities';
+import { strRemoveSymbols, strToLinks } from '../../../../../../utilities/methods/StringUtilities';
 import { CardStyle } from '../../../../../../utilities/styles/Card.style';
 import { RobotDetailControlModeTypeEnum } from '../commands/RobotDetailCommands.enum';
+import DialogNote from './DialogNote';
 import { RobotDetailGeneralInterface } from './RobotDetailGeneral.interface';
 import { RobotDetailGeneralStyle } from './RobotDetailGeneral.style';
 
@@ -18,6 +20,8 @@ const RobotDetailGeneral: FC<RobotDetailGeneralInterface> = (props) => {
 	const { t } = useTranslation(['ROBOTS', 'TOOLTIPS']);
 	const classes = RobotDetailGeneralStyle();
 	const cardClasses = CardStyle();
+
+	const [open, setOpen] = useState(false);
 
 	const common = 'CONTENT.DETAIL.GENERAL';
 
@@ -39,7 +43,7 @@ const RobotDetailGeneral: FC<RobotDetailGeneralInterface> = (props) => {
 				<Typography
 					variant="caption"
 					color="textSecondary"
-					className={classes.sGeneralItemBlock}>
+					className={classes.sGridItemBlock}>
 					{t(`${common}.STATUS.LABEL`)}
 				</Typography>
 				<Status active={robotTwins.robotState.isReady.value}>
@@ -52,7 +56,7 @@ const RobotDetailGeneral: FC<RobotDetailGeneralInterface> = (props) => {
 				<Typography
 					variant="caption"
 					color="textSecondary"
-					className={classes.sGeneralItemBlock}>
+					className={classes.sGridItemBlock}>
 					{t(`${common}.CONTROL_MODE`)}
 				</Typography>
 				<Status
@@ -62,11 +66,11 @@ const RobotDetailGeneral: FC<RobotDetailGeneralInterface> = (props) => {
 					{strRemoveSymbols(robotTwins.controlMode.value)}
 				</Status>
 			</Grid>
-			<Grid item xs={12} sm={6} md={4} lg={2} className={classes.sGeneralLastRowItem}>
+			<Grid item xs={12} sm={6} md={4} lg={2} className={classes.sGridLastRowItem}>
 				<Typography
 					variant="caption"
 					color="textSecondary"
-					className={classes.sGeneralItemBlock}>
+					className={classes.sGridItemBlock}>
 					{t(`${common}.ACCEPT_ORDERS.LABEL`)}
 				</Typography>
 				<Status active={!!robotTwins.site.acceptOrders}>
@@ -87,7 +91,7 @@ const RobotDetailGeneral: FC<RobotDetailGeneralInterface> = (props) => {
 				<Typography variant="caption" color="textSecondary">
 					{t(`${common}.MISSION`)}
 				</Typography>
-				<Typography variant="body1" className={classes.sGeneralItemFlex}>
+				<Typography variant="body1" className={classes.sGridItemFlex}>
 					{robotTwins.mission.status || AppConfigService.AppOptions.common.none}
 					{robotTwins.mission.description && (
 						<Tooltip
@@ -100,13 +104,30 @@ const RobotDetailGeneral: FC<RobotDetailGeneralInterface> = (props) => {
 									</CardContent>
 								</Card>
 							}>
-							<InfoOutlined
-								fontSize="small"
-								className={classes.sGeneralItemInfoIcon}
-							/>
+							<InfoOutlined fontSize="small" className={classes.sGridItemInfoIcon} />
 						</Tooltip>
 					)}
 				</Typography>
+			</Grid>
+			<Grid item xs={12} sm={6} md={8} lg={6} className={classes.sNoteGrid}>
+				<Typography variant="caption" color="textSecondary">
+					{t(`${common}.NOTE.TITLE`)}
+					<IconButton
+						className={classes.sNoteEditIconButton}
+						onClick={() => setOpen(true)}>
+						<Edit color="primary" className={classes.sNoteEditIcon} />
+					</IconButton>
+				</Typography>
+				<Typography
+					variant="body1"
+					dangerouslySetInnerHTML={{
+						__html:
+							(robotTwins.robot.note && strToLinks(robotTwins.robot.note)) ||
+							AppConfigService.AppOptions.common.none
+					}}
+					className={classes.sNote}
+				/>
+				{open && <DialogNote open={open} setOpen={setOpen} note={robotTwins.robot.note} />}
 			</Grid>
 		</Grid>
 	);