+++ src/app/screens/business/sites/content/configuration/notifications/SiteNotifications.tsx
@@ -1,5 +1,14 @@
-import { Box, Card, CardContent, IconButton, List, Tooltip, Typography } from '@material-ui/core';
-import AddCircleIcon from '@material-ui/icons/AddCircle';
+import {
+	Box,
+	Card,
+	CardContent,
+	Grid,
+	IconButton,
+	List,
+	Tooltip,
+	Typography
+} from '@material-ui/core';
+import { AddCircle } from '@material-ui/icons';
 import { FC, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
@@ -47,39 +56,41 @@ const SiteNotifications: FC<SiteNotificationsInterface> = (props) => {
 		return () => window.clearInterval(intervalId);
 	}, [dispatch, cSiteId]);
 
-	return site.notifications?.content ? (
-		<Card square elevation={1} className={classes.sNotificationCard}>
-			<CardContent className={cardClasses.sCardContent4}>
-				<Typography variant="h6" className={classes.sTitle}>
-					{t(`${common}.TITLE`)}
-				</Typography>
-				<Typography variant="body2" color="textSecondary" className={classes.sExcerpt}>
-					{t(`${common}.EXCERPT`)}
-				</Typography>
+	return site.notifications?.content?.data.length ? (
+		<Grid item xs={12} md={6}>
+			<Card square elevation={1} className={classes.sCard}>
+				<CardContent className={cardClasses.sCardContent4}>
+					<Typography variant="h6" className={classes.sTitle}>
+						{t(`${common}.TITLE`)}
+					</Typography>
+					<Typography variant="body2" color="textSecondary" className={classes.sExcerpt}>
+						{t(`${common}.EXCERPT`)}
+					</Typography>
 
-				<Box className={classes.sCreate}>
-					<Tooltip
-						placement="top"
-						title={String(t('TOOLTIPS:NOTIFICATION.ADD'))}
-						onClick={() => setOpen(true)}>
-						<IconButton edge="end">
-							<AddCircleIcon color="primary" />
-						</IconButton>
-					</Tooltip>
-					<DialogCreateEditNotification
-						type={SiteNotificationsCreateEditTypeEnum.CREATE}
-						open={open}
-						setOpen={setOpen}
-					/>
-				</Box>
+					<Box className={classes.sCreate}>
+						<Tooltip
+							placement="left"
+							title={String(t('TOOLTIPS:NOTIFICATION.ADD'))}
+							onClick={() => setOpen(true)}>
+							<IconButton edge="end">
+								<AddCircle color="primary" />
+							</IconButton>
+						</Tooltip>
+						<DialogCreateEditNotification
+							type={SiteNotificationsCreateEditTypeEnum.CREATE}
+							open={open}
+							setOpen={setOpen}
+						/>
+					</Box>
 
-				<List disablePadding>
-					{site.notifications.content.data.map((notification, index) => (
-						<SiteNotification key={notification.id} site={site} index={index} />
-					))}
-				</List>
-			</CardContent>
-		</Card>
+					<List disablePadding>
+						{site.notifications.content.data.map((notification, index) => (
+							<SiteNotification key={notification.id} site={site} index={index} />
+						))}
+					</List>
+				</CardContent>
+			</Card>
+		</Grid>
 	) : null;
 };
 export default SiteNotifications;