+++ src/app/slices/business/sites/Site.slice.interface.tsx
@@ -8,6 +8,9 @@ export interface SliceSiteInterface {
 	acceptOrders: {
 		loading: boolean;
 	};
+	siteRobotConfig: {
+		loading: boolean;
+	};
 	notifications: {
 		loader: boolean;
 		loading: boolean;